Adds bacalhau support
Adds a destination node for bacalhau that is able to submit jobs to a
local requester node, using the orchestrator API.

Each instance of the destination node is configured with the location of
the jobstore, a local directory containing .yaml files describing jobs.
It should also be provided with the job name (one of the .yaml files)
so that it might be configured with the following

```
jobstore: /tmp/jobstore
job: process
```

and will load /tmp/jobstore/process.yaml as the job specification.  Once
loaded the job specification can be templated using handlebars syntax,
so that {{key}} will be replaced with values from the 'key' field in the
input message.